Caseref changed to long (#11)
* CaseRef changed to long

* change test to use random caseRef

+++ src/main/java/uk/gov/ons/census/action/model/dto/PrintFileDto.java
@@ -8,7 +8,7 @@ public class PrintFileDto {
   private String qid;
   private String uacWales;
   private String qidWales;
-  private int caseRef;
+  private long caseRef;
   private String title;
   private String forename;
   private String surname;
